# Conflicts:
#	src/main/java/org/verdictdb/core/connection/SparkDataTypeConverter.java


Former-commit-id: e436ae26560ba5a05a8131d31054abb5c5803abf

+++ src/main/java/org/verdictdb/core/connection/SparkDataTypeConverter.java
@@ -32,7 +32,7 @@ public class SparkDataTypeConverter {
     } else if (type.equals("IntegerType")) {
       return INTEGER;
     } else if (type.equals("LongType")) {
-      return BIGINT;
+      return INTEGER;
     } else if (type.equals("ShortType")) {
       return SMALLINT;
     } else if (type.equals("ArrayType")) {
